requirements.txt
# Core trading and data libraries
ccxt>=4.1.0                # Cryptocurrency exchange integration
pandas>=2.0.0              # Data manipulation
numpy>=1.24.0              # Numerical computations
yfinance>=0.2.28           # Stock market data
python-binance>=1.0.19     # Binance API (optional alternative)

# Technical analysis
ta>=0.11.0                 # Technical analysis indicators
pandas-ta>=0.3.14b         # Additional technical indicators

# WebSocket for real-time data
websockets>=12.0           # WebSocket client
aiohttp>=3.9.0             # Async HTTP client

# Configuration and environment
python-dotenv>=1.0.0       # Environment variables
pyyaml>=6.0                # YAML configuration

# Utilities
requests>=2.31.0           # HTTP requests
colorama>=0.4.6            # Colored terminal output
python-dateutil>=2.8.2     # Date utilities

# Machine Learning for adaptive learning system
scikit-learn>=1.3.0        # ML models and preprocessing
scipy>=1.11.0              # Scientific computing (used by scikit-learn)
joblib>=1.3.0              # Model persistence

# Telegram notifications
python-telegram-bot==20.7  # Telegram bot API for notifications
